API security completed

+++ server/routes/about.js
@@ -6,9 +6,24 @@ const dbo = require("../db/conn");
 const ObjectId = require("mongodb").ObjectId;
 
 // This section will help you get a list of all the records.
-router.get("/", function (req, res) {
+router.get("/", async function (req, res) {
+  const header = req.headers["authorization"]
+  if (!header) {
+    res.sendStatus(403)
+    console.log("no auth")
+    return;
+  }
+  const token = header.split(" ")[1]
+  
   let db_connect = dbo.getDb();
-
+  const tokenResults = await db_connect.collection("passcode").find({token: token}).toArray()
+  if (tokenResults.length != 1) {
+    res.sendStatus(403)
+    console.log("invalid token")
+    return;
+  }
+  
+  
   db_connect
     .collection("about")
     .find({})
@@ -20,8 +35,23 @@ router.get("/", function (req, res) {
 });
 
 //post a new record
-router.post("/", function (req, response) {
+router.post("/", async function (req, response) {
+  const header = req.headers["authorization"]
+  if (!header) {
+    res.sendStatus(403)
+    console.log("no auth")
+    return;
+  }
+  const token = header.split(" ")[1]
+  
   let db_connect = dbo.getDb();
+  const tokenResults = await db_connect.collection("passcode").find({token: token}).toArray()
+  if (tokenResults.length != 1) {
+    res.sendStatus(403)
+    console.log("invalid token")
+    return;
+  }
+
   let myobj = {
     name: req.body.name,
     text: req.body.text,
@@ -37,8 +67,23 @@ router.post("/", function (req, response) {
 });
 
 //update a new record
-router.put("/:id", function (req, response) {
+router.put("/:id", async function (req, response) {
+  const header = req.headers["authorization"]
+  if (!header) {
+    res.sendStatus(403)
+    console.log("no auth")
+    return;
+  }
+  const token = header.split(" ")[1]
+  
   let db_connect = dbo.getDb();
+  const tokenResults = await db_connect.collection("passcode").find({token: token}).toArray()
+  if (tokenResults.length != 1) {
+    res.sendStatus(403)
+    console.log("invalid token")
+    return;
+  }
+
   let myquery = { _id: ObjectId(req.params.id) };
   let newvalues = {
     $set: {
@@ -60,8 +105,23 @@ router.put("/:id", function (req, response) {
 });
 
 //delete a record
-router.delete("/:id", (req, response) => {
+router.delete("/:id", async (req, response) => {
+  const header = req.headers["authorization"]
+  if (!header) {
+    res.sendStatus(403)
+    console.log("no auth")
+    return;
+  }
+  const token = header.split(" ")[1]
+  
   let db_connect = dbo.getDb();
+  const tokenResults = await db_connect.collection("passcode").find({token: token}).toArray()
+  if (tokenResults.length != 1) {
+    res.sendStatus(403)
+    console.log("invalid token")
+    return;
+  }
+  
   let myquery = { _id: ObjectId(req.params.id) };
   db_connect.collection("about").deleteOne(myquery, function (err, obj) {
     if (err) {
