+++ server/routes/home.js
@@ -6,22 +6,50 @@ const dbo = require("../db/conn");
 const ObjectId = require("mongodb").ObjectId;
 
 // This section will help you get a list of all the records.
-router.get("/", function (req, res) {
+router.get("/", async function (req, res) {
+  const header = req.headers["authorization"]
+  if (!header) {
+    res.sendStatus(403)
+    console.log("no auth")
+    return;
+  }
+  const token = header.split(" ")[1]
+  
   let db_connect = dbo.getDb();
+  const tokenResults = await db_connect.collection("passcode").find({token: token}).toArray()
+  if (tokenResults.length != 1) {
+    res.sendStatus(403)
+    console.log("invalid token")
+    return;
+  }
 
   db_connect
     .collection("home")
     .find({})
     .toArray()
     .then((response) => {
-      console.log(response);
+      // console.log(response);
       res.json(response);
     });
 });
 
 //post a new record
-router.post("/", function (req, response) {
+router.post("/", async function (req, response) {
+  const header = req.headers["authorization"]
+  if (!header) {
+    res.sendStatus(403)
+    console.log("no auth")
+    return;
+  }
+  const token = header.split(" ")[1]
+  
   let db_connect = dbo.getDb();
+  const tokenResults = await db_connect.collection("passcode").find({token: token}).toArray()
+  if (tokenResults.length != 1) {
+    res.sendStatus(403)
+    console.log("invalid token")
+    return;
+  }
   let myobj = {
     header: req.body.header,
     text: req.body.text,
@@ -40,8 +68,22 @@ router.post("/", function (req, response) {
 });
 
 //update a record
-router.put("/:id", function (req, response) {
+router.put("/:id", async function (req, response) {
+  const header = req.headers["authorization"]
+  if (!header) {
+    res.sendStatus(403)
+    console.log("no auth")
+    return;
+  }
+  const token = header.split(" ")[1]
+  
   let db_connect = dbo.getDb();
+  const tokenResults = await db_connect.collection("passcode").find({token: token}).toArray()
+  if (tokenResults.length != 1) {
+    res.sendStatus(403)
+    console.log("invalid token")
+    return;
+  }
   let myquery = { _id: ObjectId(req.params.id) };
   let newvalues = {
     $set: {
@@ -64,8 +106,22 @@ router.put("/:id", function (req, response) {
 });
 
 //delete a record
-router.delete("/:id", (req, response) => {
+router.delete("/:id", async (req, response) => {
+  const header = req.headers["authorization"]
+  if (!header) {
+    res.sendStatus(403)
+    console.log("no auth")
+    return;
+  }
+  const token = header.split(" ")[1]
+  
   let db_connect = dbo.getDb();
+  const tokenResults = await db_connect.collection("passcode").find({token: token}).toArray()
+  if (tokenResults.length != 1) {
+    res.sendStatus(403)
+    console.log("invalid token")
+    return;
+  }
   let myquery = { _id: new ObjectId(req.params.id) };
   db_connect.collection("home").deleteOne(myquery, function (err, obj) {
     if (err) {
